+++ src/store.ts
@@ -2,7 +2,7 @@ import { get, writable } from 'svelte/store';
 import { NavMode } from './components/chartUtils';
 import DataSet, { DataGroup } from './data/DataSet';
 import deriveLinkDefaults, { getDirectLinkImpl } from './deriveLinkDefaults';
-import { getFormSelections } from './components/dialogs/formSelections';
+import FormSelections from './components/dialogs/formSelections';
 
 declare const __VERSION__: string;
 
@@ -17,6 +17,18 @@ export const expandedDataGroups = writable([defaults.group]);
 export const isShowingPoints = writable(defaults.showPoints);
 export const initialViewport = writable(defaults.viewport);
 export const navMode = writable(NavMode.autofit);
+
+export function getFormSelections() {
+  try {
+    if (sessionStorage.getItem('form')) {
+      return JSON.parse(sessionStorage.getItem('form')!) as FormSelections;
+    }
+    return new FormSelections();
+  } catch {
+    sessionStorage.removeItem('form');
+    return new FormSelections();
+  }
+}
 
 export const formSelections = writable(getFormSelections());
 formSelections.subscribe((val) => {
